Added failing tests for failed retries jobs.
In particular, when a job was failing due to retries, it wasn't incrementing the
'failed' or 'failures' stats for the associated queue. This also led to a bug when
trying to cancel jobs due to failed retries.

This is also related to a bug where the 'failures' count could become negative --
probably the longest-standing and until now unreplicated bug.